+++ Configuration/TypoScript/Library/GridElements/setup.ts
@@ -1,0 +1,72 @@
+lib {
+  plugin {
+    tx_t3vcontent {
+      gridElements {
+        columnLayoutGridElement {
+          2Columns {
+            container = TEXT
+            container {
+              value = {$plugin.tx_t3vcontent.gridElements.columnLayoutGridElement.2Columns.container}
+            }
+          }
+
+          3Columns {
+            container = TEXT
+            container {
+              value = {$plugin.tx_t3vcontent.gridElements.columnLayoutGridElement.3Columns.container}
+            }
+          }
+        }
+
+        accordionGridElement {
+          container = TEXT
+          container {
+            value = {$plugin.tx_t3vcontent.gridElements.accordionGridElement.container}
+          }
+        }
+
+        accordGridElement {
+          container = TEXT
+          container {
+            value = {$plugin.tx_t3vcontent.gridElements.accordGridElement.container}
+          }
+        }
+
+        documentSectionGridElement {
+          container = TEXT
+          container {
+            value = {$plugin.tx_t3vcontent.gridElements.documentSectionGridElement.container}
+          }
+        }
+
+        panelGridElement {
+          container = TEXT
+          container {
+            value = {$plugin.tx_t3vcontent.gridElements.panelGridElement.container}
+          }
+        }
+
+        timelineGridElement {
+          container = TEXT
+          container {
+            value = {$plugin.tx_t3vcontent.gridElements.timelineGridElement.container}
+          }
+        }
+
+        eventGridElement {
+          container = TEXT
+          container {
+            value = {$plugin.tx_t3vcontent.gridElements.eventGridElement.container}
+          }
+        }
+
+        viewportGridElement {
+          container = TEXT
+          container {
+            value = {$plugin.tx_t3vcontent.gridElements.viewportGridElement.container}
+          }
+        }
+      }
+    }
+  }
+}