+++ Configuration/TypoScript/Library/constants.ts
@@ -1,3 +1,3 @@
 # === Includes ===
 
-<INCLUDE_TYPOSCRIPT: source="FILE: ./Defaults/constants.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/constants.ts">