+++ Configuration/TypoScript/Library/setup.ts
@@ -1,3 +1,3 @@
 # === Includes ===
 
-<INCLUDE_TYPOSCRIPT: source="FILE: ./Defaults/setup.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/setup.ts">