+++ ext_emconf.php
@@ -13,15 +13,15 @@ $EM_CONF[$_EXTKEY] = [
   'clearCacheOnLoad' => false,
   'constraints' => [
     'depends' => [
-      'typo3' => '8.7.0-9.5.99',
-      'fluid_styled_content' => '8.7.0-9.5.99',
+      'typo3' => '9.5.0-10.4.99',
+      'fluid_styled_content' => '9.5.0-10.4.99',
       'gridelements' => '',
       'replacecontent' => '',
       'vhs' => '',
       't3v_core' => ''
     ],
     'conflicts' => [
-      'css_styled_content' => '8.7.0-9.5.99'
+      'css_styled_content' => '9.5.0-10.4.99'
     ],
     'suggests' => []
   ],