+++ CHANGELOG.md
@@ -3,6 +3,12 @@ CHANGELOG
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+----------
+
+* Updated translations
+* Updated icons
+
 7.1.0
 -----