+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Added `Conditionally` Grid Element
 * Improved `ViewColumnViewHelper`
 * Added `ExtensionService`, introduced `strict` and `fallback` mode (breaking change)
 * Updated translations