+++ CHANGELOG.md
@@ -3,6 +3,15 @@ CHANGELOG
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+----------
+
+* Updated TypoScript configuration / structure
+* Updated Composer configuration
+* Updated Travis CI and AppVeyor configuration
+* Updated dependencies
+* Cleaned up
+
 4.11.0
 ------