+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 
 * Updated translations
 * Updated Travis CI configuration
+* Cleaned up
 
 7.3.0
 -----